Refactored out the StringValue parsing into its own class
